Move add volume functionallity from tests/utils to libvmi/storage
